Move DisableGraphAttribute to io.crate namespace
