+++ server/src/main/java/io/crate/lucene/DisableGraphAttribute.java
@@ -17,11 +17,11 @@
  * under the License.
  */
 
-package org.apache.lucene.analysis.miscellaneous;
+package io.crate.lucene;
 
 import org.apache.lucene.analysis.TokenStream;
-import org.apache.lucene.util.Attribute;
 import org.apache.lucene.analysis.tokenattributes.PositionLengthAttribute;
+import org.apache.lucene.util.Attribute;
 
 /**
  * This attribute can be used to indicate that the {@link PositionLengthAttribute}
